Adds listing of transfers to next-gw-notification
