+++ src/Slackbot.Net.Extensions.FplBot/Abstractions/ITransfersByGameWeek.cs
@@ -1,0 +1,9 @@
+using System.Threading.Tasks;
+
+namespace Slackbot.Net.Extensions.FplBot.Abstractions
+{
+    internal interface ITransfersByGameWeek
+    {
+        Task<string> GetTransfersByGameweek(int? gw);
+    }
+}
