+++ src/Slackbot.Net.Extensions.FplBot/SlackBotBuilderExtensions.cs
@@ -40,6 +40,7 @@ namespace Slackbot.Net.Abstractions.Hosting
         private static void AddCommon(this ISlackbotWorkerBuilder builder)
         {
             builder.Services.AddSingleton<ICaptainsByGameWeek, CaptainsByGameWeek>();
+            builder.Services.AddSingleton<ITransfersByGameWeek,TransfersByGameWeek>();
             builder.Services.AddSingleton<IChipsPlayed, ChipsPlayed>();
             builder.Services.AddSingleton<ITeamValue, TeamValue>();
             builder.Services.AddSingleton<IMessageHelper, MessageHelper>();
